% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnmrQuant1D.R
\name{RnmrQuant1D}
\alias{RnmrQuant1D}
\title{RnmrQuant1D}
\description{
Package dedicated to 1D proton NMR quantification, including peak fitting and based on external calibration using standard spectra.
}
\details{
This package was initially developed as part of an ANR project on wine authenticity (\href{https://anr.fr/Project-ANR-21-CE21-0014}{ANR-21-CE21-0014}). However, it is generic enough to be used on other biological and/or food matrices. This involves the implementation of an analytical protocol allowing quantification from an external standard (see references).\cr\cr
This package provides all the methods needed: \cr 
\enumerate{
   \item to calculate different entities such as i) integration of patterns related to the targeted compounds of interest, ii) signal-to-noise ratio of motifs, iii) absolute quantifications of the targeted compounds,
   \item to calculate PULCON factors and verify using a quality control sample,
   \item  to exploit the results i) by producing a complete workbook gathering all the calculated data but also information needed for tracing, ii) by visualizing the spectra zones related to each targeted compound including the result of the deconvolution and the identification of patterns.
}
\cr
To work, this package needs a number of input files:\cr
\enumerate{
   \item a set of spectra concerning the samples. There may be for the same sample several repetitions acquired according to one or more sequences (zgpr, noesy)
   \item a set of spectra concerning quality control (QC) and quantification standards (QS). Similarly, there may be several repetitions acquired according to one or more sequences (zgpr, noesy)
   \item a file describing the samples by linking them to spectral data
   \item a file describing the QC and the QS samples in terms of compounds and their concentration, called "calibration profile"
   \item a file containing the parameters for deconvolution and quantification, zone by zone, called "quantification profile"
}
\cr
The user can refer to the \href{https://github.com/djacob65/RnmrQuant1D/wiki/}{online tutorial} for more details on the structure of the input files, and on the different steps to perform the quantifications, from data preparation to final results.\cr
\cr 
It was implemented based on the object-oriented programming using the R6 class.
}
\references{
Guillaume Leleu et al. (2024) Development of a standardized method for metabolite analysis by NMR to assess wine authenticity, IVES Conference Series, OIV, DOI:10.58233/IMGGSIME

Teipel et al. (2020) Application of 1H Nuclear Magnetic Resonance Spectroscopy as Spirit Drinks Screener for Quality and Authenticity Control, Food 9, 1355; doi:10.3390/foods9101355
}
\author{
Daniel Jacob - (C) INRAE 2025
}
\section{Super class}{
\code{RnmrQuant1D::internalClass} -> \code{RnmrQuant1D}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{RAWDIR}}{Path of the sample spectra directory (required)}

\item{\code{QSDIR}}{Path of the QC-QS spectra directory (required)}

\item{\code{TMPDIR}}{Output path for logs (required)}

\item{\code{RDATADIR}}{Output path for RData files (required)}

\item{\code{SAMPLES}}{The sample table (required)}

\item{\code{PROFILE}}{The quantification profile obtained by the \href{#method-RnmrQuant1D-readProfile}{\code{readProfile()}} method (required).}

\item{\code{CALIBRATION}}{The calibration profile as a data.frame (required)}

\item{\code{SEQUENCE}}{The instrument sequence (required)}

\item{\code{FIELD}}{The instrument Field (recommended for tracing)}

\item{\code{TYPE}}{The sample type (recommended for tracing)}

\item{\code{OUTTYPE}}{specifies the output type for plotly graphs and data tables. The value can be "html", "png", "jpeg", or "svg" (default = "html")}

\item{\code{FDILfield}}{Name of the dilution factor column in the sample table (default = "F_dilution")}

\item{\code{QCtype}}{Identifier of the Quality Control (QC) in the calibration profile (1st column - default = "QC")}

\item{\code{QStype}}{Identifier of the Quantification Standards (QS) in the calibration profile (1st column - default = "QS")}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RnmrQuant1D-new}{\code{RnmrQuant1D$new()}}
\item \href{#method-RnmrQuant1D-check_samples}{\code{RnmrQuant1D$check_samples()}}
\item \href{#method-RnmrQuant1D-check_calibration}{\code{RnmrQuant1D$check_calibration()}}
\item \href{#method-RnmrQuant1D-check_profile}{\code{RnmrQuant1D$check_profile()}}
\item \href{#method-RnmrQuant1D-check_outdir}{\code{RnmrQuant1D$check_outdir()}}
\item \href{#method-RnmrQuant1D-check_all}{\code{RnmrQuant1D$check_all()}}
\item \href{#method-RnmrQuant1D-get_samples_table}{\code{RnmrQuant1D$get_samples_table()}}
\item \href{#method-RnmrQuant1D-readProfile}{\code{RnmrQuant1D$readProfile()}}
\item \href{#method-RnmrQuant1D-get_compound_table}{\code{RnmrQuant1D$get_compound_table()}}
\item \href{#method-RnmrQuant1D-get_response_factors}{\code{RnmrQuant1D$get_response_factors()}}
\item \href{#method-RnmrQuant1D-get_factor_table}{\code{RnmrQuant1D$get_factor_table()}}
\item \href{#method-RnmrQuant1D-get_QC_estimation}{\code{RnmrQuant1D$get_QC_estimation()}}
\item \href{#method-RnmrQuant1D-plot_QC_estimation}{\code{RnmrQuant1D$plot_QC_estimation()}}
\item \href{#method-RnmrQuant1D-proc_Integrals}{\code{RnmrQuant1D$proc_Integrals()}}
\item \href{#method-RnmrQuant1D-get_Matrix_Integrals}{\code{RnmrQuant1D$get_Matrix_Integrals()}}
\item \href{#method-RnmrQuant1D-get_Matrix_SNR}{\code{RnmrQuant1D$get_Matrix_SNR()}}
\item \href{#method-RnmrQuant1D-get_Matrix_CV}{\code{RnmrQuant1D$get_Matrix_CV()}}
\item \href{#method-RnmrQuant1D-save_Matrices}{\code{RnmrQuant1D$save_Matrices()}}
\item \href{#method-RnmrQuant1D-view_spectra}{\code{RnmrQuant1D$view_spectra()}}
\item \href{#method-RnmrQuant1D-proc_fPULCON}{\code{RnmrQuant1D$proc_fPULCON()}}
\item \href{#method-RnmrQuant1D-proc_Quantification}{\code{RnmrQuant1D$proc_Quantification()}}
\item \href{#method-RnmrQuant1D-get_output_results}{\code{RnmrQuant1D$get_output_results()}}
\item \href{#method-RnmrQuant1D-get_spectra_data}{\code{RnmrQuant1D$get_spectra_data()}}
\item \href{#method-RnmrQuant1D-save_Results}{\code{RnmrQuant1D$save_Results()}}
\item \href{#method-RnmrQuant1D-plot_spectra}{\code{RnmrQuant1D$plot_spectra()}}
\item \href{#method-RnmrQuant1D-displayWidget}{\code{RnmrQuant1D$displayWidget()}}
\item \href{#method-RnmrQuant1D-saveWidgetFix}{\code{RnmrQuant1D$saveWidgetFix()}}
\item \href{#method-RnmrQuant1D-displayTable}{\code{RnmrQuant1D$displayTable()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="RnmrQuant1D" data-topic="internalClass" data-id="get_Matrix_Stats"><a href='../../RnmrQuant1D/html/internalClass.html#method-internalClass-get_Matrix_Stats'><code>RnmrQuant1D::internalClass$get_Matrix_Stats()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RnmrQuant1D" data-topic="internalClass" data-id="reorderProfile"><a href='../../RnmrQuant1D/html/internalClass.html#method-internalClass-reorderProfile'><code>RnmrQuant1D::internalClass$reorderProfile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RnmrQuant1D" data-topic="internalClass" data-id="saveProfile"><a href='../../RnmrQuant1D/html/internalClass.html#method-internalClass-saveProfile'><code>RnmrQuant1D::internalClass$saveProfile()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RnmrQuant1D-new"></a>}}
\if{latex}{\out{\hypertarget{method-RnmrQuant1D-new}{}}}
\subsection{Method \code{new()}}{
Class Constructor - Initializes internal parameters for preprocessing and deconvolution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RnmrQuant1D$new()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{self}}{the RnmrQuant1D instance}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a RnmrQuant1D instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RnmrQuant1D-check_samples"></a>}}
\if{latex}{\out{\hypertarget{method-RnmrQuant1D-check_samples}{}}}
\subsection{Method \code{check_samples()}}{
Checks if the sample table format is right and if sample names correspond to spectra under the RAWDIR directory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RnmrQuant1D$check_samples(verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{Display the success message if TRUE}

\item{\code{self}}{The RnmrQuant1D instance}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RnmrQuant1D-check_calibration"></a>}}
\if{latex}{\out{\hypertarget{method-RnmrQuant1D-check_calibration}{}}}
\subsection{Method \code{check_calibration()}}{
Checks if the calibration profile is OK
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RnmrQuant1D$check_calibration(
  QC = NULL,
  QS = NULL,
  sequence = NULL,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{QC}}{If specified, this is the name of the spectra directory for quality control (QC) (default is NULL)}

\item{\code{QS}}{If specified, this is the name of the spectra directory for qualification standards (QS) (default is NULL)}

\item{\code{sequence}}{If specified, this is the chosen sequence for testing. By default (NULL), the sequence in the RnmrQuant1D instance will be chosen.}

\item{\code{verbose}}{Display the success message if TRUE}

\item{\code{self}}{The RnmrQuant1D instance}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RnmrQuant1D-check_profile"></a>}}
\if{latex}{\out{\hypertarget{method-RnmrQuant1D-check_profile}{}}}
\subsection{Method \code{check_profile()}}{
Checks if the quantification profile is OK
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RnmrQuant1D$check_profile(zones = NULL, verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{zones}}{specifies the zone(s) to be checked}

\item{\code{verbose}}{Display the success message if TRUE}

\item{\code{self}}{The RnmrQuant1D instance}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RnmrQuant1D-check_outdir"></a>}}
\if{latex}{\out{\hypertarget{method-RnmrQuant1D-check_outdir}{}}}
\subsection{Method \code{check_outdir()}}{
Checks if output directories (TMPDIR & RDATADIR) are OK
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RnmrQuant1D$check_outdir(verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{Display the success message if TRUE}

\item{\code{self}}{The RnmrQuant1D instance}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RnmrQuant1D-check_all"></a>}}
\if{latex}{\out{\hypertarget{method-RnmrQuant1D-check_all}{}}}
\subsection{Method \code{check_all()}}{
Checks many things : directories, profiles (quantification & calibration), ...
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RnmrQuant1D$check_all(verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{Display the success message if TRUE}

\item{\code{self}}{The RnmrQuant1D instance}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RnmrQuant1D-get_samples_table"></a>}}
\if{latex}{\out{\hypertarget{method-RnmrQuant1D-get_samples_table}{}}}
\subsection{Method \code{get_samples_table()}}{
Generates the sample table for the corresponding sequence.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RnmrQuant1D$get_samples_table(sequence = NULL, infos = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sequence}}{If specified, take samples defined for this NMR sequence. Otherwise, takes samples for the one defined in the instance.}

\item{\code{infos}}{If TRUE, add some additional columns such as the TMSP width at half height and other acquisition parameters.}

\item{\code{self}}{The RnmrQuant1D instance}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the sample table
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RnmrQuant1D-readProfile"></a>}}
\if{latex}{\out{\hypertarget{method-RnmrQuant1D-readProfile}{}}}
\subsection{Method \code{readProfile()}}{
Reads the quantification profile. See the \href{https://github.com/djacob65/RnmrQuant1D/wiki/}{wiki page} for having more details.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RnmrQuant1D$readProfile(PROFILE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PROFILE}}{the quantification profile filename}

\item{\code{self}}{The RnmrQuant1D instance}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the qualification profile object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RnmrQuant1D-get_compound_table"></a>}}
\if{latex}{\out{\hypertarget{method-RnmrQuant1D-get_compound_table}{}}}
\subsection{Method \code{get_compound_table()}}{
Gets the compounds by zone from the quantification profile.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RnmrQuant1D$get_compound_table(zones = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{zones}}{the quantification zones}

\item{\code{self}}{The RnmrQuant1D instance}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a dataframe with 2 columns : zones & compounds
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RnmrQuant1D-get_response_factors"></a>}}
\if{latex}{\out{\hypertarget{method-RnmrQuant1D-get_response_factors}{}}}
\subsection{Method \code{get_response_factors()}}{
Calculates the response factors according to the sample type ('QC' or 'QS').
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RnmrQuant1D$get_response_factors(
  sampletype,
  samplename,
  thresfP = 5,
  deconv = TRUE,
  verbose = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sampletype}}{the sample type : 'QC' for Quality Control, or 'QS' for Quantification Standards.}

\item{\code{samplename}}{a valid sample name under the spectra directory (QSDIR)}

\item{\code{thresfP}}{defines the threshold of the CV (in percentage) of the response factor below which the spectrum (repetition) will not be taken into consideration.}

\item{\code{deconv}}{defines whether the integration is calculated by deconvolution or by the Simpson approach.}

\item{\code{verbose}}{if TRUE, some messages are displayed}

\item{\code{self}}{The RnmrQuant1D instance}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a response factor object, i.e a list with some attributes: fPUL, fP, fR, MC, INTG, fK
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RnmrQuant1D-get_factor_table"></a>}}
\if{latex}{\out{\hypertarget{method-RnmrQuant1D-get_factor_table}{}}}
\subsection{Method \code{get_factor_table()}}{
Returns the factor table according to the sample type ('QC' or 'QS').
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RnmrQuant1D$get_factor_table(QS)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{QS}}{a response factor object returned by \href{#method-RnmrQuant1D-get_response_factors}{\code{get_response_factors()}}.}

\item{\code{self}}{The RnmrQuant1D instance}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the factor table (matrix)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RnmrQuant1D-get_QC_estimation"></a>}}
\if{latex}{\out{\hypertarget{method-RnmrQuant1D-get_QC_estimation}{}}}
\subsection{Method \code{get_QC_estimation()}}{
Estimation of the compound concentrations (QC) based on the PULCON factor (QS)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RnmrQuant1D$get_QC_estimation(QC, QS, merge = TRUE, verbose = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{QC}}{a response factor object for a QC type returned by \href{#method-RnmrQuant1D-get_response_factors}{\code{get_response_factors()}}.}

\item{\code{QS}}{a response factor object for a QS type returned by \href{#method-RnmrQuant1D-get_response_factors}{\code{get_response_factors()}}.}

\item{\code{merge}}{If TRUE, merge compounds that have the same names but with a different number at end, e.g. "citrate1", "citrate2" will be merged under "citrate".}

\item{\code{verbose}}{If TRUE, some messages are displayed}

\item{\code{self}}{The RnmrQuant1D instance}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the estimation table (matrix)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RnmrQuant1D-plot_QC_estimation"></a>}}
\if{latex}{\out{\hypertarget{method-RnmrQuant1D-plot_QC_estimation}{}}}
\subsection{Method \code{plot_QC_estimation()}}{
Plots the estimation of the compound concentrations (QC) based on the PULCON factor (QS)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RnmrQuant1D$plot_QC_estimation(QCest)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{QCest}}{the estimation table returned by \href{#method-RnmrQuant1D-get_QC_estimation}{\code{get_QC_estimation()}}.}

\item{\code{self}}{The RnmrQuant1D instance}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a \href{https://plotly.com/r/}{plotly} graph
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RnmrQuant1D-proc_Integrals"></a>}}
\if{latex}{\out{\hypertarget{method-RnmrQuant1D-proc_Integrals}{}}}
\subsection{Method \code{proc_Integrals()}}{
Performs the integration of all compounds defined by a pattern included in the specified zone(s).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RnmrQuant1D$proc_Integrals(zones, ncpu = 2, verbose = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{zones}}{The selected zones for integration.}

\item{\code{ncpu}}{The number of processors (cores) used for parallel computing.}

\item{\code{verbose}}{if TRUE, some messages are displayed}

\item{\code{self}}{The RnmrQuant1D instance}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The result will be stored in the RnmrQuant1D instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RnmrQuant1D-get_Matrix_Integrals"></a>}}
\if{latex}{\out{\hypertarget{method-RnmrQuant1D-get_Matrix_Integrals}{}}}
\subsection{Method \code{get_Matrix_Integrals()}}{
Gets the integration matrix from the RnmrQuant1D instance, computed by \href{#method-RnmrQuant1D-proc_Integrals}{\code{proc_Integrals()}}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RnmrQuant1D$get_Matrix_Integrals()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{self}}{The RnmrQuant1D instance}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the integration matrix
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RnmrQuant1D-get_Matrix_SNR"></a>}}
\if{latex}{\out{\hypertarget{method-RnmrQuant1D-get_Matrix_SNR}{}}}
\subsection{Method \code{get_Matrix_SNR()}}{
Gets the SNR matrix from the RnmrQuant1D instance, computed by \href{#method-RnmrQuant1D-proc_Integrals}{\code{proc_Integrals()}}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RnmrQuant1D$get_Matrix_SNR()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{self}}{The RnmrQuant1D instance}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the SNR matrix
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RnmrQuant1D-get_Matrix_CV"></a>}}
\if{latex}{\out{\hypertarget{method-RnmrQuant1D-get_Matrix_CV}{}}}
\subsection{Method \code{get_Matrix_CV()}}{
Gets the CV matrix from the integration matrix if samples have repetitions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RnmrQuant1D$get_Matrix_CV(MatInt = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{MatInt}}{The integration matrix. If NULL, it will be fetched from the the RnmrQuant1D instance. See \href{#method-RnmrQuant1D-get_Matrix_Integrals}{\code{get_Matrix_Integrals()}}.}

\item{\code{self}}{The RnmrQuant1D instance}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the CV matrix
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RnmrQuant1D-save_Matrices"></a>}}
\if{latex}{\out{\hypertarget{method-RnmrQuant1D-save_Matrices}{}}}
\subsection{Method \code{save_Matrices()}}{
Saves the result tables (samples, integration, SNR) computed by \href{#method-RnmrQuant1D-proc_Integrals}{\code{proc_Integrals()}} and other information for tracing (quantification profile, R environment, ...) into a workbook (\href{https://docs.fileformat.com/spreadsheet/xlsx/}{XLSX format}).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RnmrQuant1D$save_Matrices(file, filelist = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{the path to the output file}

\item{\code{filelist}}{a simple list containing two elements, just for tracing : 1) SAMPLEFILE, the path to the sample file, 2) PROFILE, the path to the quantification profile.}

\item{\code{self}}{The RnmrQuant1D instance}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RnmrQuant1D-view_spectra"></a>}}
\if{latex}{\out{\hypertarget{method-RnmrQuant1D-view_spectra}{}}}
\subsection{Method \code{view_spectra()}}{
Plots the ppm range defined by 'RnmrQuant1D$ppm_range', the latter being positioned by the zone(s) chosen when calculating the integrations by \href{#method-RnmrQuant1D-proc_Integrals}{\code{proc_Integrals()}}. The graph superimposes several spectra, namely the original spectrum, the model spectrum resulting from the deconvolution, as well as all the patterns corresponding to the compounds identified according to the profile in the preselected zone(s).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RnmrQuant1D$view_spectra(
  id,
  plotmodel = TRUE,
  plotTrueSpec = TRUE,
  plotzones = TRUE,
  tags = "none",
  showlegend = TRUE,
  legendhoriz = FALSE,
  showgrid = TRUE,
  title = NULL,
  colspecs = NULL,
  colcpmds = NULL,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Either the order number (numeric) of the spectrum in the sample table or the samplecode (string), i.e the 2nd column in the sample table.}

\item{\code{plotmodel}}{If TRUE, plot the model based on the deconvolution.}

\item{\code{plotTrueSpec}}{If TRUE, plot the original spectrum without local baseline corrections.}

\item{\code{plotzones}}{If TRUE, add a semi-transparent rectangle superimposed on each fit zone.}

\item{\code{tags}}{add an arrow annotation on each peak resulting from the deconvolution. Among the possible values: 1) 'id' adds the number of the compound to which the peak belongs, 2) 'name' adds the name of the compound to which the peak belongs, 'auto' adds either the number if multiple zones, or the name of the compound, 3) 'none' adds no arrow annotation.}

\item{\code{showlegend}}{If TRUE, show the legend}

\item{\code{legendhoriz}}{Put the legend at the bottom (TRUE) or at the left (FALSE) of the graph}

\item{\code{showgrid}}{If TRUE, show the grid on the background}

\item{\code{title}}{If NULL, the title will be the sample code. Otherwise, it will be the specified string. If you don't want a title, enter an empty string.}

\item{\code{colspecs}}{specifies the array of colors for the original spectrum and the model. Only the two first colors will be used.}

\item{\code{colcpmds}}{specifies the array of colors for the compounds (i.e. patterns). If the number of colors specified is not sufficient, the colors will be repeated.}

\item{\code{verbose}}{If TRUE, The peak list will be displayed}

\item{\code{self}}{The RnmrQuant1D instance}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a \href{https://plotly.com/r/}{plotly} graph
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RnmrQuant1D-proc_fPULCON"></a>}}
\if{latex}{\out{\hypertarget{method-RnmrQuant1D-proc_fPULCON}{}}}
\subsection{Method \code{proc_fPULCON()}}{
Computes the response factor for the QS sample.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RnmrQuant1D$proc_fPULCON(QSname, thresfP = 5, deconv = TRUE, verbose = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{QSname}}{a valid QS sample name under the spectra directory (QSDIR)}

\item{\code{thresfP}}{defines the threshold of the CV (in percentage) of the response factor below which the spectrum (repetition) will not be taken into consideration.}

\item{\code{deconv}}{defines whether the integration is calculated by deconvolution or by the Simpson approach.}

\item{\code{verbose}}{if TRUE, some messages are displayed}

\item{\code{self}}{The RnmrQuant1D instance}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
he resulting object (fP) will be stored in the RnmrQuant1D instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RnmrQuant1D-proc_Quantification"></a>}}
\if{latex}{\out{\hypertarget{method-RnmrQuant1D-proc_Quantification}{}}}
\subsection{Method \code{proc_Quantification()}}{
Computes for each sample the concentration  of targeted compounds  based on the quantification profile.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RnmrQuant1D$proc_Quantification(
  cmpdlist = NULL,
  zones = NULL,
  ncpu = 2,
  reset = FALSE,
  CR = FALSE,
  verbose = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cmpdlist}}{Sets the list of targeted compounds. It can be null and in this case, it will be the list of zones that will be taken into consideration.}

\item{\code{zones}}{In case the list of targeted compounds is null, then the targeted compounds will be defined by the patterns included in the indicated zones. If the list of zones is null, as well as that of the targeted compounds (cmpdlist), then takes into consideration all the zones defined in the quantification profile.}

\item{\code{ncpu}}{The number of processors (cores) used for parallel computing.}

\item{\code{reset}}{Indicates whether to delete all existing RData files or to proceed by accumulation. In this latter case, if some samples had already been processed then they will not be processed again.}

\item{\code{CR}}{Indicates whether to add a Carriage Return for a more presentable display (case with Jupyter notebooks)}

\item{\code{verbose}}{if TRUE, some messages are displayed}

\item{\code{self}}{The RnmrQuant1D instance}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The results will be stored under the \code{"RDATADIR"} directory as RData files.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RnmrQuant1D-get_output_results"></a>}}
\if{latex}{\out{\hypertarget{method-RnmrQuant1D-get_output_results}{}}}
\subsection{Method \code{get_output_results()}}{
Retrieves from RData files on disk generated by \href{#method-RnmrQuant1D-proc_Quantification}{\code{proc_Quantification()}}, the result tables within a list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RnmrQuant1D$get_output_results()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{self}}{The RnmrQuant1D instance}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the result tables within a list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RnmrQuant1D-get_spectra_data"></a>}}
\if{latex}{\out{\hypertarget{method-RnmrQuant1D-get_spectra_data}{}}}
\subsection{Method \code{get_spectra_data()}}{
Retrieves from RData files generated on disk by \href{#method-RnmrQuant1D-proc_Quantification}{\code{proc_Quantification()}}, the spectra and the peaklist within lists in the RnmrQuant1D instance respectively named specList and res.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RnmrQuant1D$get_spectra_data()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{self}}{The RnmrQuant1D instance}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RnmrQuant1D-save_Results"></a>}}
\if{latex}{\out{\hypertarget{method-RnmrQuant1D-save_Results}{}}}
\subsection{Method \code{save_Results()}}{
Saves the result tables (samples, integration, SNR, quantifications) computed by \href{#method-RnmrQuant1D-proc_Quantification}{\code{proc_Quantification()}} and other information for tracing (calibration profile, quantification profile, R environment, ...) into a workbook (\href{https://docs.fileformat.com/spreadsheet/xlsx/}{XLSX format}).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RnmrQuant1D$save_Results(file, filelist = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{the path to the output file}

\item{\code{filelist}}{a simple list containing three elements, just for tracing : 1) SAMPLEFILE, the path to the sample file, 2) PROFILE, the path to the quantification profile, 3) CALIBRATION, the path to the calibration profile}

\item{\code{self}}{The RnmrQuant1D instance}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RnmrQuant1D-plot_spectra"></a>}}
\if{latex}{\out{\hypertarget{method-RnmrQuant1D-plot_spectra}{}}}
\subsection{Method \code{plot_spectra()}}{
Visualizes the compound 'compound' for spectrum 'id', where id is the order number of the spectrum in the sample table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RnmrQuant1D$plot_spectra(id, compound, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Either the order number (numeric) of the spectrum in the sample table or the samplecode (string), i.e the 2nd column in the sample table.}

\item{\code{compound}}{The compound name taken from quantpars$cmpdlist in the RnmrQuant1D instance.}

\item{\code{...}}{Arguments to be passed to the \href{#method-RnmrQuant1D-view_spectra}{\code{view_spectra()}} method.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a \href{https://plotly.com/r/}{plotly} graph
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RnmrQuant1D-displayWidget"></a>}}
\if{latex}{\out{\hypertarget{method-RnmrQuant1D-displayWidget}{}}}
\subsection{Method \code{displayWidget()}}{
Displays a widget (e.g. a plotly graph) - Depending on the OutType attribute in the RnmrQuant1D instance, returns a \href{https://plotly.com/r/}{plotly} graph (html) or displays the image (png,svg, ...)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RnmrQuant1D$displayWidget(widget, tmpdir = "tmp", width = "auto", height = 400)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{widget}}{a plotly graph}

\item{\code{tmpdir}}{the directory to temporarily save the image before displaying it (png,svg, ...)}

\item{\code{width}}{the width of the output image}

\item{\code{height}}{the height of the output image}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
depending on the OutType attribute in the RnmrQuant1D instance, returns a \href{https://plotly.com/r/}{plotly} graph (html) or display the image (png,svg, ...)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RnmrQuant1D-saveWidgetFix"></a>}}
\if{latex}{\out{\hypertarget{method-RnmrQuant1D-saveWidgetFix}{}}}
\subsection{Method \code{saveWidgetFix()}}{
Saves a widget (e.g. a plotly graph) into a file - A wrapper to saveWidget which compensates for arguable BUG in htmlwidgets::saveWidget which requires `file` to be in current working directory.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RnmrQuant1D$saveWidgetFix(widget, file, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{widget}}{a plotly graph}

\item{\code{file}}{the path of the output filename}

\item{\code{...}}{Arguments to be passed to htmlwidgets::saveWidget}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RnmrQuant1D-displayTable"></a>}}
\if{latex}{\out{\hypertarget{method-RnmrQuant1D-displayTable}{}}}
\subsection{Method \code{displayTable()}}{
Beautifies a matrix using the DT package before displaying it
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RnmrQuant1D$displayTable(M, nbdec = 2, tmpdir = "tmp")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{M}}{the matrix to be embellished with the DT package}

\item{\code{nbdec}}{the number of decimal places for floating point numbers in the output table}

\item{\code{tmpdir}}{the directory to temporarily save the image before displaying it}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Depending on the OutType attribute in the RnmrQuant1D instance, returns a \href{https://plotly.com/r/}{plotly} graph (html) or displays the image (png,svg, ...)
}
}
}
